+++ project_test/project/foo/failure.py
@@ -1,0 +1,9 @@
+"""
+This file contain a syntax error for some test purpose.
+
+Don't correct it.
+"""
+from rstview.registry import rstview_registry
+from django.utils.translation import ugettext_lazy
+
+rstview_registry.update({