+++ project_test/tests/test_004_view.py
@@ -37,8 +37,6 @@ def test_render_basic(settings, client, urlname, output_filename, writer):
 
     assert response.content == attempted
 
-    #assert 1 == 42
-
 
 def test_empty_doc_path(settings, client):
     """Empty doc_path class attribute raise an exception"""